agents/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "praisonaiagents"
version = "0.0.36"
description = "Praison AI agents for completing complex tasks with Self Reflection Agents"
authors = [
    { name="Mervin Praison" }
]
dependencies = [
    "pydantic",
    "rich",
    "openai"
]

[project.optional-dependencies]
memory = [
    "chromadb>=0.6.0"
]

knowledge = [
    "mem0ai>=0.1.0",
    "chromadb==0.5.23",
    "markitdown"
]

# Combined features
all = [
    "praisonaiagents[memory]",
    "praisonaiagents[knowledge]"
]